complete file upload and store image

+++ controllers/admin.js
@@ -1,7 +1,11 @@
 const Product = require("../models/product");
 const mongoose = require("mongoose");
 const {validationResult}  = require("express-validator/check");
+const fileHelper = require("../util/file");
+const multer = require("multer");
 exports.getAddProduct = (req, res, next) => {  
+  console.log(1);
+  console.log(req.session.csrfToken)
   res.render("admin/edit-product", {
     pageTitle: "Add Product",
     path: "/admin/add-product" ,
@@ -17,47 +21,59 @@ exports.getAddProduct = (req, res, next) => {
   });
 };
 
-
 exports.postAddProduct = async (req, res, next) => {
-  const title = req.body.title.trim() ; 
-  const imageUrl = req.body.imageUrl; 
-  const price = +req.body.price;
-  const description = req.body.description.trim();
-  console.log(title);
-  const errors = validationResult(req);
-  if(!errors.isEmpty()){
-    const errorsArr = errors.array();
-    return res.status(422).render("admin/edit-product", {
-      pageTitle: "Add Product",
-      path: "/admin/add-product" ,
-      editMode : false, 
-      errorMessage : errorsArr.map(({msg}) => msg),
-      product : {
-        title, 
-        imageUrl,
-        price ,
-        description
-      },
-      validationErrors : errors.array()
-    })
-  }
-  try {     
-    let product = new Product({      
-      title,
-      imageUrl,
-      price,
-      description,
-      userId: req.user,
-    });
-    await product.save();  
-    return res.redirect("/");
-  } catch (err) {   
-    console.log(err);    
-    const error = new Error(err);
-    error.httpStatusCode = 500 ; 
-    return next(error);
-  }
- 
+    const title = req.body.title.trim() ; 
+    const image = req.file; 
+    const price = +req.body.price;
+    const description = req.body.description.trim();
+    console.log(image);
+    if(!image){
+      return res.status(422).render("admin/edit-product", {
+        pageTitle: "Add Product",
+        path: "/admin/add-product" ,
+        editMode : false, 
+        errorMessage : "Attached file is not an image",
+        product : {
+          title,          
+          price ,
+          description
+        },
+        validationErrors : []
+      })
+    }
+    const errors = validationResult(req);
+    if(!errors.isEmpty()){
+      const errorsArr = errors.array();
+      return res.status(422).render("admin/edit-product", {
+        pageTitle: "Add Product",
+        path: "/admin/add-product" ,
+        editMode : false, 
+        errorMessage : errorsArr.map(({msg}) => msg),
+        product : {
+          title,           
+          price ,
+          description
+        },
+        validationErrors : errors.array()
+      })
+    }
+    const imageUrl = image.path ;
+    try {           
+      let product = new Product({      
+        title : title ,
+        imageUrl : imageUrl ,
+        price : price ,
+        description : description,
+        userId: req.user,
+      });   
+      await product.save() ;
+      return res.redirect("/");
+    } catch (err) {      
+      console.log(err);
+      const error = new Error(err);
+      error.httpStatusCode = 500; 
+      next(error)
+    }
 };
 
 exports.getEditProduct = async (req, res, next) => {  
@@ -85,7 +101,7 @@ exports.getEditProduct = async (req, res, next) => {
 exports.postEditProduct = async (req, res, next) => {  
   const productId = req.body.productId;
   const title = req.body.title ;
-  const imageUrl = req.body.imageUrl;
+  const image = req.file;
   const price = req.body.price ;
   const description = req.body.description;
   const errors = validationResult(req);
@@ -100,7 +116,6 @@ exports.postEditProduct = async (req, res, next) => {
       product : {
         _id : productId,
         title, 
-        imageUrl,
         price ,
         description
       },
@@ -110,7 +125,10 @@ exports.postEditProduct = async (req, res, next) => {
   try {    
     let product = await Product.findById(productId);    
     product.title = title;
-    product.imageUrl = imageUrl;
+    if(image){
+      fileHelper.deleteFile(product.imageUrl);
+      product.imageUrl = image.path;
+    }
     product.description = description;
     product.price = price;
     await product.save();
@@ -148,7 +166,13 @@ exports.getProducts = async (req, res, next) => {
 exports.postDeleteProduct = async (req, res, next) => {
   try {
     const id = req.body.id ;   
-    let result = await Product.deleteOne({_id : id , userId : req.user._id });   
+    let product=  await Product.findById(id);
+    if(!product){
+      return next(new Error("No Product found"));
+    }   
+    await Product.deleteOne({_id : id , userId : req.user._id });
+    fileHelper.deleteFile(product.imageUrl);
+    console.log("DELETE SUCCESS")
     return res.redirect("/admin/products")
   } catch (err) {
     console.log(err);
